moved services out into a different folder [Jerry Chang]
